third_party/blink/web_tests/storage/indexeddb/deletedatabase-not-blocked-expected.txt

Test that deleteDatabase is not blocked when connections close in on versionchange callback

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "deletedatabase-not-blocked.html"
indexedDB.open(dbname)


openOnSuccess():
h = event.target.result
indexedDB.deleteDatabase(dbname)


onVersionChange():
PASS event.target.version is 1
PASS event.oldVersion is 1
PASS event.newVersion is null
h.close()


deleteDatabaseOnSuccess():
PASS blocked event was not fired
PASS successfullyParsed is true

TEST COMPLETE
